+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/informers/cache/Controller.java
@@ -91,16 +91,17 @@ public class Controller<T extends HasMetadata, L extends KubernetesResourceList<
 
     // Starts one daemon thread for resync
     this.resyncExecutor = Executors.newSingleThreadScheduledExecutor();
-    initReflector();
+    this.reflector = new Reflector<>(apiTypeClass, listerWatcher, queue, operationContext, fullResyncPeriod);
   }
 
   public void run() {
     log.info("informer#Controller: ready to run resync and reflector runnable");
-
     // Start the resync runnable
     if (fullResyncPeriod > 0) {
       ResyncRunnable resyncRunnable = new ResyncRunnable(queue, resyncFunc);
-      resyncFuture = resyncExecutor.scheduleAtFixedRate(resyncRunnable, fullResyncPeriod, fullResyncPeriod, TimeUnit.MILLISECONDS);
+      if(!resyncExecutor.isShutdown()) {
+        resyncFuture = resyncExecutor.scheduleAtFixedRate(resyncRunnable, fullResyncPeriod, fullResyncPeriod, TimeUnit.MILLISECONDS);
+      }
     } else {
       log.info("informer#Controller: resync skipped due to 0 full resync period");
     }
@@ -123,6 +124,10 @@ public class Controller<T extends HasMetadata, L extends KubernetesResourceList<
     synchronized (this) {
       reflector.stop();
       reflectExecutor.shutdown();
+      if (resyncFuture != null) {
+        resyncFuture.cancel(true);
+      }
+      resyncExecutor.shutdown();
     }
   }
 
@@ -139,9 +144,6 @@ public class Controller<T extends HasMetadata, L extends KubernetesResourceList<
    * @return latest resource version
    */
   public String lastSyncResourceVersion() {
-    if (reflector == null) {
-      return "";
-    }
     return reflector.getLastSyncResourceVersion();
   }
 
@@ -167,7 +169,11 @@ public class Controller<T extends HasMetadata, L extends KubernetesResourceList<
     }
   }
 
-  private void initReflector() {
-      reflector = new Reflector<>(apiTypeClass, listerWatcher, queue, operationContext, fullResyncPeriod);
+  ScheduledExecutorService getReflectExecutor() {
+    return this.reflectExecutor;
+  }
+
+  ScheduledExecutorService getResyncExecutor() {
+    return this.resyncExecutor;
   }
 }